Relay setup properly

+++ server/data/loaders/UserLoader.js
@@ -33,7 +33,11 @@ function updateUser(user) {
 }
 
 function deleteUser(id) {
-  return User.destroy({where: {id: id}});
+  return User.find({where: {id: id}}).then(function (user) {
+    if(!user) return null;
+    User.destroy({where: {id: id}});
+    return user;
+  });
 }
 
 export {
